+++ ChangeLog.md
@@ -1,5 +1,5 @@
 # Changelog - UltimateDRM
 
-- July 20 2025: Added module checksums  
+- July 20 2025: Added module checksums, periodic checksum comparisons    
 - July 19 2025: Added code signing checks, PIMPL idiom for safe static .lib usage  
 - July 18 2025: First commit. Added section remapping.